+++ src/codegen/mod.rs
@@ -33,7 +33,7 @@ pub fn codegen_function_definition(function: &ast::FunctionDefinition) -> asm::F
 pub fn codegen_expression(expression: &ast::Expression) -> asm::Operand {
     let ast::Expression { kind, span: _ } = expression;
     match kind {
-        ast::ExpressionKind::Constant(literal) => asm::Operand::Immediate(literal.value),
+        ast::ExpressionKind::Constant(literal) => asm::Operand::Immediate(literal.value as isize),
         ast::ExpressionKind::Call(_) => todo!(),
     }
 }
